Discard changes to test.py and create test_introspection.py. Implement other PR recommendations.
